Add files via upload
this is simple implementation of BFS path finding and still not complete as i want.
the main code obviously is the algorithm itself and it imports walls and settings form other 2 files, 
grid editor is another program i wrote as a handy tool for different projects involving grid/board kind a things 
one problem i couldn’t solve was filtering the shortest path. i couldn’t visually only show the shortest path, but shortest path is there included all the paths from starting point to the destination. i would be happy if some one direct me about it, or any thought in general.
i wrote the same code in python which is easier in some areas and harder in another but there i could visually separate and display the shortest path.